+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - [#515](https://github.com/okta/okta-auth-js/pull/515) Removes `token.value` field
 - [#540](https://github.com/okta/okta-auth-js/pull/540) Locks `tokenManager.expireEarlySeconds` option with the default value (30s) for non-dev environment
+### Other
+
+- [#675](https://github.com/okta/okta-auth-js/pull/675) Calls updateAuthState in the constructor to trigger update authState to the latest state
+  
 ## 4.9.0
 
 ### Bug Fixes
@@ -31,7 +35,6 @@
 - [#648](https://github.com/okta/okta-auth-js/pull/648) Updates widget to 5.4.2
 - [#653](https://github.com/okta/okta-auth-js/pull/653) Removes isLoginRedirect check in oidc logic
 - [#661](https://github.com/okta/okta-auth-js/pull/661) Upgrades node-cache to 5.1.2
-  
 ## 4.7.2
 
 ### Bug Fixes